Refactor subscribe tests into worker-internals.test.ts and add smoke test docs
Split worker tests: subscribe tests now import the function directly
instead of extracting it via createWorkerFactory side-effect capture.
Added SMOKETESTS.md documenting manual example verification steps.
